Move checking of expired tasks to home page and fix home page message when all tasks expired
