+++ src/app/pages/home/containers/home-page.component.html
@@ -1,12 +1,6 @@
 <ion-header>
   <ion-toolbar color="secondary">
-    <ion-button
-      class="logo-small"
-      ion-button
-      menuToggle
-      slot="start"
-      (click)="showCredits()"
-    >
+    <ion-button class="logo-small" ion-button menuToggle slot="start" (click)="showCredits()">
       <img class="logo-small" src="assets/imgs/radar-logo.svg" alt="Logo" />
     </ion-button>
     <ion-button ion-button menuToggle slot="end" (click)="openSettingsPage()">
@@ -17,48 +11,26 @@
 </ion-header>
 
 <ion-content>
-  <div
-    class="size-dynamic"
-    [class.hidden]="showCalendar"
-    [@moveProgress]="showCalendar"
-  >
-    <task-progress
-      *ngIf="tasksProgress | async"
-      [progress]="tasksProgress | async"
-      [forceComplete]="showCompleted"
-      [noTasksToday]="(tasksProgress | async)?.numberOfTasks == 0"
-      [streakDays]="streakDays"
-      [isStreakShown]="isStreakShown | async"
-    >
+  <div class="size-dynamic" [class.hidden]="showCalendar" [@moveProgress]="showCalendar">
+    <task-progress *ngIf="tasksProgress | async" [progress]="tasksProgress | async" [forceComplete]="showCompleted"
+      [noTasksToday]="(tasksProgress | async)?.numberOfTasks == 0" [streakDays]="streakDays"
+      [isStreakShown]="isStreakShown | async">
     </task-progress>
   </div>
 
   <div class="bottom-components" [class.bottom-components-up]="showCalendar">
     <div class="clinical-container" *ngIf="showMiscTasksButton | async">
-      <ion-fab
-        left
-        bottom
-        class="clinical-button"
-        [class.clinical-button-bottom]="showCalendar"
-      >
+      <ion-fab left bottom class="clinical-button" [class.clinical-button-bottom]="showCalendar">
         <ion-fab-button ion-fab class="fab">
           <ion-icon name="add"></ion-icon>
         </ion-fab-button>
         <ion-fab-list side="top">
-          <ion-fab-button
-            *ngIf="hasClinicalTasks | async"
-            class="clinical-icon"
-            ion-fab
-            (click)="openClinicalTasksPage()"
-          >
+          <ion-fab-button *ngIf="hasClinicalTasks | async" class="clinical-icon" ion-fab
+            (click)="openClinicalTasksPage()">
             <img src="assets/imgs/clinical.png" class="icon-clinical" />
           </ion-fab-button>
-          <ion-fab-button
-            *ngIf="hasOnDemandTasks | async"
-            class="clinical-icon"
-            ion-fab
-            (click)="openOnDemandTasksPage()"
-          >
+          <ion-fab-button *ngIf="hasOnDemandTasks | async" class="clinical-icon" ion-fab
+            (click)="openOnDemandTasksPage()">
             <img src="{{ onDemandIcon | async }}" class="icon-clinical" />
           </ion-fab-button>
         </ion-fab-list>
@@ -66,78 +38,40 @@
     </div>
 
     <div #tickerBar class="size-static">
-      <ticker-bar
-        [task]="nextTask"
-        [isNow]="taskIsNow"
-        [timeToNextMilli]="timeToNextTask"
-        [showAffirmation]="showCompleted"
-        [showSyncNeeded]="showSyncNeeded"
-        [noTasksToday]="(tasksProgress | async)?.numberOfTasks == 0"
-      ></ticker-bar>
+      <ticker-bar [task]="nextTask" [isNow]="taskIsNow" [timeToNextMilli]="timeToNextTask"
+        [showAffirmation]="showCompleted" [showSyncNeeded]="showSyncNeeded" [showExpired]="showExpired"
+        [noTasksToday]="(tasksProgress | async)?.numberOfTasks == 0"></ticker-bar>
     </div>
 
     <div class="size-static">
-      <task-info
-        *ngIf="nextTask && (isTaskInfoShown | async)"
-        [task]="nextTask"
-        [isNow]="taskIsNow"
-        [progress]="tasksProgress | async"
-        [expanded]="!showCalendar"
-        (click)="displayTaskCalendar()"
-      ></task-info>
+      <task-info *ngIf="nextTask && (isTaskInfoShown | async)" [task]="nextTask" [isNow]="taskIsNow"
+        [progress]="tasksProgress | async" [expanded]="!showCalendar" (click)="displayTaskCalendar()"></task-info>
     </div>
     <div #taskCalendar class="taskCalendar" [@displayCalendar]="showCalendar">
-      <task-calendar
-        *ngIf="
+      <task-calendar *ngIf="
           (sortedTasks | async)?.size &&
           (isTaskCalendarTaskNameShown | async) !== null
-        "
-        [show]="showCalendar"
-        (task)="startQuestionnaire($event)"
-        [tasks]="sortedTasks | async"
-        [currentDate]="currentDate"
-        [isTaskNameShown]="isTaskCalendarTaskNameShown | async"
-      >
+        " [show]="showCalendar" (task)="startQuestionnaire($event)" [tasks]="sortedTasks | async"
+        [currentDate]="currentDate" [isTaskNameShown]="isTaskCalendarTaskNameShown | async">
       </task-calendar>
     </div>
 
     <div class="size-static footer">
-      <div
-        *ngIf="showCompleted && (studyPortalReturnUrl | async)"
-        class="ion-padding"
-      >
-        <ion-button
-          ion-button
-          class="bt bt--full"
-          round
-          full
-          (click)="openStudyPortal()"
-        >
+      <div *ngIf="showCompleted && (studyPortalReturnUrl | async)" class="ion-padding">
+        <ion-button ion-button class="bt bt--full" round full (click)="openStudyPortal()">
           {{ portalReturnText | async }}
         </ion-button>
       </div>
-      <div
-        *ngIf="getIsStartButtonShown() && (sortedTasks | async)?.size"
-        class="ion-padding"
-      >
-        <ion-button
-          ion-button
-          class="bt bt--full"
-          round
-          full
-          (click)="startQuestionnaire(null)"
-        >
+      <div *ngIf="getIsStartButtonShown() && (sortedTasks | async)?.size" class="ion-padding">
+        <ion-button ion-button class="bt bt--full" round full (click)="startQuestionnaire(null)">
           {{ 'BTN_START' | translate }}
         </ion-button>
       </div>
-      <div
-        *ngIf="
-          getIsLoadingSpinnerShown() ||
+      <div *ngIf="
+      getIsLoadingSpinnerShown() ||
           (!(sortedTasks | async)?.size &&
             (tasksProgress | async)?.numberOfTasks)
-        "
-        class="spinner-center"
-      >
+        " class="spinner-center">
         <ion-spinner name="dots"></ion-spinner>
       </div>
     </div>
